Factor out txn.rs from mod.rs in persistence api
